data: Move selected clear reasons
Selected-change state now has separate file-change payload storage, while
the main store still owns hunk state, snapshots, kind markers, and
cleared-selection refusal markers.

The project lacks a distinct owner for cleared-selection markers. That
makes command refusal behavior share a module with unrelated persisted hunk
state.

This commit continues splitting selected-change persistence by moving
clear-reason markers and refusal helpers into
data.selected_change.clear_reasons and updating source callers.

The next commit will update tests and architecture guards to name the
clear-reason owner.
