+++ src/git_stage_batch/commands/skip.py
@@ -32,6 +32,8 @@ from ..data.selected_change.store import (
     SelectedChangeKind,
     get_selected_change_file_path,
     read_selected_change_kind,
+)
+from ..data.selected_change.clear_reasons import (
     refuse_bare_action_after_auto_advance_disabled,
     refuse_bare_action_after_file_list,
 )
