+++ src/git_stage_batch/commands/status.py
@@ -39,8 +39,10 @@ from ..data.progress import format_id_range
 from ..data.selected_change.lifecycle import clear_selected_change_state_files
 from ..data.selected_change.store import (
     SelectedChangeKind,
-    mark_selected_change_cleared_by_stale_batch_selection,
     read_selected_change_kind,
+)
+from ..data.selected_change.clear_reasons import (
+    mark_selected_change_cleared_by_stale_batch_selection,
 )
 from ..data.selected_change.file_changes import (
     load_selected_binary_file,
